+++ src/Classes/DraggableComponent.ts
@@ -1,6 +1,6 @@
 import Component from "./Component";
 import Renderer from "./Renderer";
-import { Vector2 } from "./types";
+import { Vector2 } from "../utils/types";
 
 class DraggableComponent extends Component {
   private initalPosition: Vector2 = { x: 0, y: 0 };
