+++ src/State/State.ts
@@ -1,0 +1,3 @@
+class State {}
+
+export default State;
